Added MainMenu view with Settings
Created new issue for the MainMenu and Settings view.
The InputManager we have is added to the scene,but it doesn't work tho.
I added more methods to the Patient script according to the UML diagram
Added Player component, fixed retypo  and removed script changes not related to this issue